proxycfg: unexport agent cache test fixtures
These test helpers are only used within this package, and the many test helpers in this
package pollute the godoc and make it hard to see the real interface.
